rename seal module to groth16
